backend/pyproject.toml
[tool.poetry]
name = "backend"
version = "0.1.0"
description = ""
authors = ["Md. Nazrul Islam Khan <nazrulkhan.dk@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
Django = "^4.2.1"
pillow = "^9.5.0"
djangorestframework = "^3.14.0"
boto3 = "^1.26.142"
whitenoise = {extras = ["brotli"], version = "^6.4.0"}
django-filter = "^22.1"
django-storages = "^1.13.2"
django-extensions = "^3.2.1"
django-cors-headers = "^3.14.0"
attrs = "^22.2.0"
psycopg = "^3.1.9"
redis = "^4.5.5"
gunicorn = "^20.1.0"
python-dotenv = "^1.0.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
